chore: restructure components folder
